refactor: http 관련 폴더 변경
