+++ tomcat/src/main/java/org/apache/coyote/MimeType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote;
 
 import org.apache.coyote.util.FileExtension;
