+++ tomcat/src/main/java/org/apache/coyote/controller/FrontController.java
@@ -7,8 +7,8 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.http11.HttpStatusCode;
-import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.HttpStatusCode;
+import org.apache.coyote.MimeType;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.ResponseHeader;
