+++ tomcat/src/main/java/org/apache/coyote/controller/LoginController.java
@@ -5,11 +5,11 @@ import com.techcourse.model.User;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.catalina.Manager;
+import org.apache.coyote.HttpStatusCode;
+import org.apache.coyote.MimeType;
 import org.apache.coyote.Session;
 import org.apache.coyote.SessionManager;
 import org.apache.coyote.http11.HttpCookie;
-import org.apache.coyote.http11.HttpStatusCode;
-import org.apache.coyote.http11.MimeType;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.RequestHeader;
 import org.apache.coyote.http11.response.HttpResponse;
