+++ tomcat/src/main/java/org/apache/coyote/controller/NotFoundController.java
@@ -5,8 +5,8 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.apache.coyote.http11.HttpStatusCode;
-import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.HttpStatusCode;
+import org.apache.coyote.MimeType;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.ResponseHeader;
