+++ tomcat/src/main/java/org/apache/coyote/controller/RegisterController.java
@@ -4,8 +4,8 @@ import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.http11.HttpStatusCode;
-import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.HttpStatusCode;
+import org.apache.coyote.MimeType;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.ResponseHeader;
