+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponse.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11.response;
 
 import java.util.StringJoiner;
+import org.apache.coyote.HttpStatusCode;
 import org.apache.coyote.HttpVersion;
-import org.apache.coyote.http11.HttpStatusCode;
 
 public class HttpResponse {
