+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseHeader.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.response;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.StringJoiner;
-import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.MimeType;
 
 public class ResponseHeader {
