+++ tomcat/src/test/java/org/apache/coyote/controller/LoginControllerTest.java
@@ -8,8 +8,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import org.apache.coyote.http11.HttpStatusCode;
-import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.HttpStatusCode;
+import org.apache.coyote.MimeType;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.junit.jupiter.api.BeforeEach;
