+++ tomcat/src/test/java/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -3,6 +3,8 @@ package org.apache.coyote.http11;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import org.apache.coyote.HttpStatusCode;
+import org.apache.coyote.MimeType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
