+++ tomcat/src/test/java/org/apache/coyote/http11/HttpStateCodeTest.java
@@ -2,6 +2,7 @@ package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.HttpStatusCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
