+++ tomcat/src/test/java/org/apache/coyote/http11/MimeTypeTest.java
@@ -2,6 +2,7 @@ package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.MimeType;
 import org.apache.coyote.util.FileExtension;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
