+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpRequestTest.java
@@ -7,7 +7,7 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
-import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
