+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestLineTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Map;
-import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
