+++ tomcat/src/test/java/org/apache/coyote/http11/response/StatusLineTest.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.response;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.apache.coyote.HttpVersion;
-import org.apache.coyote.http11.HttpStatusCode;
+import org.apache.coyote.HttpStatusCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
